Test/requirements.txt
# Python Dependencies for Concurrent Testing

# Selenium for browser automation (Firefox)
selenium>=4.15.0

# Supabase Python client
supabase>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Additional utilities
urllib3>=2.0.0
certifi>=2023.7.22

# Note: GeckoDriver (Firefox WebDriver) must be installed separately
# Download from: https://github.com/mozilla/geckodriver/releases
# Or install via package manager:
#   Windows: choco install selenium-gecko-driver
#   macOS: brew install geckodriver
#   Linux: sudo apt-get install firefox-geckodriver
